refactor(primitives): move TrackerPolicy, TORRENT_PEERS_LIMIT, and PrivateMode from configuration to primitives
Moves three domain primitive types from `torrust-tracker-configuration`
into `torrust-tracker-primitives` where they semantically belong.

- Add `packages/primitives/src/policy.rs` with `TrackerPolicy` struct and
  `TORRENT_PEERS_LIMIT` constant
- Add `packages/primitives/src/mode.rs` with `PrivateMode` struct
- Export new modules from `torrust-tracker-primitives` lib
- Remove types from `torrust-tracker-configuration`
- Update all import sites in `tracker-core`, `swarm-coordination-registry`,
  and `torrent-repository-benchmarking`
- Remove `torrust-tracker-configuration` dependency from
  `swarm-coordination-registry` and `torrent-repository-benchmarking`

Closes #1859
