+++ packages/primitives/src/lib.rs
@@ -5,10 +5,12 @@
 //! by the tracker server crate, but also by other crates in the Torrust
 //! ecosystem.
 pub mod announce;
+pub mod mode;
 pub mod number_of_bytes;
 pub mod pagination;
 pub mod peer;
 pub mod peer_id;
+pub mod policy;
 pub mod scrape;
 pub mod swarm_metadata;
 
@@ -16,8 +18,10 @@ use std::collections::BTreeMap;
 
 pub use announce::{AnnounceData, AnnounceEvent, AnnouncePolicy};
 use bittorrent_primitives::info_hash::InfoHash;
+pub use mode::PrivateMode;
 pub use number_of_bytes::NumberOfBytes;
 pub use peer_id::{PeerClient, PeerId};
+pub use policy::{TORRENT_PEERS_LIMIT, TrackerPolicy};
 pub use scrape::ScrapeData;
 /// Duration since the Unix Epoch.
 ///
