+++ packages/swarm-coordination-registry/src/swarm/registry.rs
@@ -5,10 +5,9 @@ use crossbeam_skiplist::SkipMap;
 use tokio::sync::Mutex;
 use torrust_clock::DurationSinceUnixEpoch;
 use torrust_clock::conv::convert_from_timestamp_to_datetime_utc;
-use torrust_tracker_configuration::TrackerPolicy;
 use torrust_tracker_primitives::pagination::Pagination;
 use torrust_tracker_primitives::swarm_metadata::{AggregateActiveSwarmMetadata, SwarmMetadata};
-use torrust_tracker_primitives::{NumberOfDownloads, NumberOfDownloadsBTreeMap, peer};
+use torrust_tracker_primitives::{NumberOfDownloads, NumberOfDownloadsBTreeMap, TrackerPolicy, peer};
 
 use crate::CoordinatorHandle;
 use crate::event::Event;
@@ -676,9 +675,8 @@ mod tests {
                 use std::sync::Arc;
 
                 use torrust_clock::DurationSinceUnixEpoch;
-                use torrust_tracker_configuration::TORRENT_PEERS_LIMIT;
                 use torrust_tracker_primitives::peer::Peer;
-                use torrust_tracker_primitives::{AnnounceEvent, NumberOfBytes};
+                use torrust_tracker_primitives::{AnnounceEvent, NumberOfBytes, TORRENT_PEERS_LIMIT};
 
                 use crate::swarm::registry::Registry;
                 use crate::swarm::registry::tests::the_swarm_repository::numeric_peer_id;
@@ -756,7 +754,7 @@ mod tests {
 
             use bittorrent_primitives::info_hash::InfoHash;
             use torrust_clock::DurationSinceUnixEpoch;
-            use torrust_tracker_configuration::TrackerPolicy;
+            use torrust_tracker_primitives::TrackerPolicy;
 
             use crate::swarm::registry::Registry;
             use crate::tests::{sample_info_hash, sample_peer};
@@ -1438,7 +1436,7 @@ mod tests {
 
             // Remove peerless torrents
 
-            let tracker_policy = torrust_tracker_configuration::TrackerPolicy {
+            let tracker_policy = torrust_tracker_primitives::TrackerPolicy {
                 remove_peerless_torrents: true,
                 ..Default::default()
             };
