+++ packages/torrent-repository-benchmarking/Cargo.toml
@@ -23,7 +23,6 @@ futures = "0"
 parking_lot = "0"
 tokio = { version = "1", features = [ "macros", "net", "rt-multi-thread", "signal", "sync" ] }
 torrust-clock = { version = "3.0.0-develop", path = "../clock" }
-torrust-tracker-configuration = { version = "3.0.0-develop", path = "../configuration" }
 torrust-tracker-primitives = { version = "3.0.0-develop", path = "../primitives" }
 
 [dev-dependencies]
