+++ packages/torrent-repository-benchmarking/src/entry/mutex_parking_lot.rs
@@ -2,9 +2,8 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 
 use torrust_clock::DurationSinceUnixEpoch;
-use torrust_tracker_configuration::TrackerPolicy;
-use torrust_tracker_primitives::peer;
 use torrust_tracker_primitives::swarm_metadata::SwarmMetadata;
+use torrust_tracker_primitives::{TrackerPolicy, peer};
 
 use super::{Entry, EntrySync};
 use crate::{EntryMutexParkingLot, EntrySingle};
