+++ packages/torrent-repository-benchmarking/src/repository/dash_map_mutex_std.rs
@@ -3,10 +3,9 @@ use std::sync::Arc;
 use bittorrent_primitives::info_hash::InfoHash;
 use dashmap::DashMap;
 use torrust_clock::DurationSinceUnixEpoch;
-use torrust_tracker_configuration::TrackerPolicy;
 use torrust_tracker_primitives::pagination::Pagination;
 use torrust_tracker_primitives::swarm_metadata::{AggregateActiveSwarmMetadata, SwarmMetadata};
-use torrust_tracker_primitives::{NumberOfDownloads, NumberOfDownloadsBTreeMap, peer};
+use torrust_tracker_primitives::{NumberOfDownloads, NumberOfDownloadsBTreeMap, TrackerPolicy, peer};
 
 use super::Repository;
 use crate::entry::peer_list::PeerList;
