+++ packages/torrent-repository-benchmarking/src/repository/rw_lock_std_mutex_tokio.rs
@@ -6,10 +6,9 @@ use bittorrent_primitives::info_hash::InfoHash;
 use futures::future::join_all;
 use futures::{Future, FutureExt};
 use torrust_clock::DurationSinceUnixEpoch;
-use torrust_tracker_configuration::TrackerPolicy;
 use torrust_tracker_primitives::pagination::Pagination;
 use torrust_tracker_primitives::swarm_metadata::{AggregateActiveSwarmMetadata, SwarmMetadata};
-use torrust_tracker_primitives::{NumberOfDownloads, NumberOfDownloadsBTreeMap, peer};
+use torrust_tracker_primitives::{NumberOfDownloads, NumberOfDownloadsBTreeMap, TrackerPolicy, peer};
 
 use super::RepositoryAsync;
 use crate::entry::peer_list::PeerList;
