+++ packages/torrent-repository-benchmarking/tests/repository/mod.rs
@@ -3,10 +3,9 @@ use std::hash::{DefaultHasher, Hash, Hasher};
 
 use bittorrent_primitives::info_hash::InfoHash;
 use rstest::{fixture, rstest};
-use torrust_tracker_configuration::TrackerPolicy;
 use torrust_tracker_primitives::pagination::Pagination;
 use torrust_tracker_primitives::swarm_metadata::SwarmMetadata;
-use torrust_tracker_primitives::{AnnounceEvent, NumberOfBytes, NumberOfDownloadsBTreeMap};
+use torrust_tracker_primitives::{AnnounceEvent, NumberOfBytes, NumberOfDownloadsBTreeMap, TrackerPolicy};
 use torrust_tracker_torrent_repository_benchmarking::EntrySingle;
 use torrust_tracker_torrent_repository_benchmarking::entry::Entry as _;
 use torrust_tracker_torrent_repository_benchmarking::repository::dash_map_mutex_std::XacrimonDashMap;
